LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/parsing/marker-supported-properties-expected.txt


FAIL Property font value 'italic small-caps 900 expanded 25px / 50px Ahem' in ::marker assert_in_array: value "italic small-caps 900 expanded 25px/normal Ahem" not in array ["italic small-caps 900 expanded 25px / 50px Ahem", "italic small-caps 900 expanded 25px/50px Ahem"]
PASS Property font-family value 'Ahem' in ::marker
PASS Property font-feature-settings value '"smcp"' in ::marker
FAIL Property font-kerning value 'none' in ::marker assert_equals: expected "none" but got "auto"
PASS Property font-size value '25px' in ::marker
FAIL Property font-size-adjust value '1' in ::marker assert_true: font-size-adjust doesn't seem to be supported in the computed style expected true got false
PASS Property font-stretch value 'expanded' in ::marker
PASS Property font-style value 'italic' in ::marker
PASS Property font-synthesis value 'none' in ::marker
PASS Property font-variant value 'small-caps' in ::marker
PASS Property font-variant-caps value 'small-caps' in ::marker
PASS Property font-variant-east-asian value 'full-width' in ::marker
PASS Property font-variant-ligatures value 'historical-ligatures' in ::marker
PASS Property font-variant-numeric value 'slashed-zero' in ::marker
PASS Property font-variant-position value 'sub' in ::marker
PASS Property font-weight value '900' in ::marker
FAIL Property line-height value '50px' in ::marker assert_equals: expected "50px" but got "normal"
FAIL Property white-space value 'nowrap' in ::marker assert_equals: expected "nowrap" but got "normal"
PASS Property color value 'rgb(0, 100, 200)' in ::marker
FAIL Property text-combine-upright value 'all' in ::marker assert_true: text-combine-upright doesn't seem to be supported in the computed style expected true got false
FAIL Property unicode-bidi value 'plaintext' in ::marker assert_equals: expected "plaintext" but got "normal"
FAIL Property direction value 'rtl' in ::marker assert_equals: expected "rtl" but got "ltr"
FAIL Property content value '"foo"' in ::marker assert_equals: expected "\"foo\"" but got ""
FAIL Property animation value '1s linear 2s infinite alternate forwards paused anim' in ::marker assert_equals: expected "1s linear 2s infinite alternate forwards paused anim" but got ""
PASS Property animation-delay value '1s' in ::marker
PASS Property animation-direction value 'alternate' in ::marker
PASS Property animation-duration value '2s' in ::marker
PASS Property animation-fill-mode value 'forwards' in ::marker
PASS Property animation-iteration-count value 'infinite' in ::marker
PASS Property animation-name value 'anim' in ::marker
PASS Property animation-play-state value 'paused' in ::marker
PASS Property animation-timing-function value 'linear' in ::marker
PASS Property transition value 'display 1s linear 2s' in ::marker
PASS Property transition-delay value '1s' in ::marker
PASS Property transition-duration value '2s' in ::marker
PASS Property transition-property value 'display' in ::marker
PASS Property transition-timing-function value 'linear' in ::marker
FAIL Property hyphens value 'none' in ::marker assert_true: hyphens doesn't seem to be supported in the computed style expected true got false
FAIL Property letter-spacing value '10px' in ::marker assert_equals: expected "10px" but got "normal"
FAIL Property line-break value 'anywhere' in ::marker assert_equals: expected "anywhere" but got "auto"
FAIL Property overflow-wrap value 'anywhere' in ::marker assert_equals: expected "anywhere" but got "normal"
FAIL Property tab-size value '10px' in ::marker assert_equals: expected "10px" but got "8"
FAIL Property text-transform value 'uppercase' in ::marker assert_equals: expected "uppercase" but got "none"
FAIL Property word-break value 'break-word' in ::marker assert_equals: expected "break-word" but got "normal"
FAIL Property word-spacing value '10px' in ::marker assert_equals: expected "10px" but got "0px"
FAIL Property text-decoration-skip-ink value 'none' in ::marker assert_equals: expected "none" but got "auto"
FAIL Property text-emphasis value 'dot rgb(0, 255, 0)' in ::marker assert_equals: expected "dot rgb(0, 255, 0)" but got ""
FAIL Property text-emphasis-color value 'rgb(0, 255, 0)' in ::marker assert_equals: expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
FAIL Property text-emphasis-position value 'under left' in ::marker assert_equals: expected "under left" but got "over right"
FAIL Property text-emphasis-style value 'dot' in ::marker assert_equals: expected "dot" but got "none"
FAIL Property text-shadow value 'rgb(0, 255, 0) 1px 2px 3px' in ::marker assert_equals: expected "rgb(0, 255, 0) 1px 2px 3px" but got "none"
PASS Property display value 'none' in ::marker
PASS Property position value 'absolute' in ::marker
PASS Property float value 'right' in ::marker
FAIL Property list-style value 'inside url('foo') decimal' in ::marker assert_equals: expected "outside none disc" but got "disc outside none"
PASS Property list-style-image value 'url('foo')' in ::marker
PASS Property list-style-position value 'inside' in ::marker
PASS Property list-style-type value 'decimal' in ::marker
target
